Fix Flask-Migrate setup
`flask db init` should only be run once, when creates `migrations/`.  This
directories contents should be stored in version control (and it is).  This
directory is used by Flask-Migrate to manage database migrations using Alembic
and SQLAlchemy. However, once it has been created, `flask db init` does not
need to be run again, and therefore should not be run by new developers when
installing their development environment.

`flask db migrate` is used to create a migration script based on the current
database schema and new changes to the models. This should be used by
developers to create a migration if and when they change a model (add, remove,
or update one or more models). However, it should not be called when a new
developer is installing their development environment.

`flask db upgrade` is used to run migration scripts against ones current
database to synchronize it with the application's current schema. In the case
of Sqlite, this command will also create the database file. New developers
should run this when installing their development enviornment.

For the above reasons, this commit changes the following section of the install
instructions in README.md

```
--> Run db migrations
$ flask db init
$ flask db migrate
```

to the following

```
--> Run db migrations
$ flask db upgrade
```

However, to allow `flask db upgrade` to work properly, this commit also removes
the call to `db.create_all()` in `backend/app.py`.  The problem was that `flask
db upgrade` loads app.py which calls `db.create_all()` which creates all the
tables of the database based on the applciation's models; and it does all of
this before it gets the chance to run the migrations; so when the migrations
ran, it would complain that the tables that the migrations where supposed to
create already existed.