Prevent normalizing EOLs on select .gitignore
The problem is that the MacOS file Icon\r\r has two carriage returns at the end of the file name. If git normalizes them we lose these characters. This commit adds a rule to .gitattributes to not normalize EOL in /.gitignore (which is generated anyway) and /.gitigonre.d/macOS.gitignore (which is where the Icon\r\r rule appears).